Organize rmf_site_editor example
