+++ crates/rmf_site_sim/diagrams/rmf_site_editor_example.mmd
@@ -4,12 +4,16 @@ sequenceDiagram
     participant robot
     participant door
 
-    request_generator->>planner: TaskRequest
-    robot->>planner: RobotPoseUpdate
-    planner->>robot: AddTrajectory
-    door->>robot: DoorState(closed)
-    robot->>robot: UpdateTracking(follow_delay)
-    robot->>door: DoorCommand(open)
-    door->>robot: DoorState(open)
-    robot->>robot: UpdateTracking(follow_delay)
-    robot->>robot: GoalReached
+    request_generator->>planner: TaskState::Active
+    planner->>robot: AssignRobotTrajectory
+    robot->>robot: Pose
+    robot->>door: DoorCommand::Open
+    robot->>robot: RobotTrajectory::hold
+    door->>door: DoorState::Opening
+    door->>robot: DoorState::Open
+    robot->>robot: RobotTrajectory::resume
+    robot->>robot: Pose
+    robot->>door: DoorCommand::Close
+    door->>door: DoorState::Closing
+    door->>door: DoorState::Closed
+    robot->>robot: TaskState::Complete
